Remove external styles
