--- styles/movies.js
@@ -1,12 +1,0 @@
-/* styles.js */
-import css from 'styled-jsx/css'
-
-export default css`
-.movies {
-  padding: 1rem;
-}
-.movies-list__directed-by {
-  display: block;
-  font-size: 1rem;
-}
-`
