--- styles/people.js
@@ -1,13 +1,0 @@
-/* styles.js */
-import css from 'styled-jsx/css'
-
-export default css`
-.people {
-  padding: 1rem;
-}
-
-.people .list h3 {
-  line-height: 1em;
-  padding: 0.5em 0;
-}
-`
